Move check_ntp_sync_status to ntp_helper to fix cross-feature import
